Refactoring functionality into core transformer, should prevent some future bugs in odd cases
